refactor: 불필요한 ProductList 컴포넌트의 래퍼 컴포넌트와 depth 제거 (#1)
