+++ src/pages/SavingsCalculatorPage/components/index.ts
@@ -1,6 +1,5 @@
 export * from './common';
-export { default as FilteredProducts } from './FilteredProducts';
 export { default as CalculationResult } from './CalculationResult';
-export { default as RecommendedProducts } from './RecommendedProducts';
 export { default as AmountInput } from './AmountInput';
 export { default as TermSelect } from './TermSelect';
+export { default as ProductList } from './ProductList';
